+++ core/docs/changelog.md
@@ -1,5 +1,8 @@
 _For details about Kurtosis' versioning scheme, as well as how to upgrade, see [the versioning & upgrading page](./versioning-and-upgrading.md)_
 
+# TBD
+* Refactor ServiceNetwork into several smaller components, and add tests for them
+
 # 1.6.4
 * Modify CI to fail the build when `ERRO` shows up, to catch bugs that may not present in the exit code
 * When a container using an IP is destroyed, release it back into the free IP address tracker's pool